기능수정: 내 플레이리스트에 노래 추가 및 삭제 기능

+++ src/main/java/project/singasong/playlistSong/repository/PlaylistSongRepository.java
@@ -12,6 +12,6 @@ import project.singasong.playlistSong.dto.PlaylistSongPagingDto;
 public interface PlaylistSongRepository {
     List<PlaylistSongPagingDto> findByPlaylistId(PlaylistSongPagingDto playlistSongPagingDto);
     Optional<PlaylistSong> findByPlaylistIdAndSongId(PlaylistSong playlistSong);
-    int create(PlaylistSong playlistSong);
+    long create(PlaylistSong playlistSong);
     int delete(Long playlistId, Long songId);
 }
